// src/types/Dog.ts
export interface Dog {
  id: string;
  name: string;
  breed: string;
  age: string;
  price: number;
  image: string;
  description: string;
  personality: string;
  origin: string;
  trainingTips: string;
  careTips: string;
  vaccinated: boolean;
  dewormed: boolean;
  microchipped: boolean;
  pedigree: boolean;
  gender: 'Mâle' | 'Femelle';
  weight: string;
  color: string;
}